# cmd/responses/user.go
package responses

// CreateUserResponse create user response
type CreateUserResponse struct {
	Username string `json:"username"`
}

// UpdateUserResponse update user response
type UpdateUserResponse struct {
	Username string `json:"username"`
}